moved function wrapper to its own header
