+++ CMakeLists.txt
@@ -32,6 +32,7 @@ set_target_properties(cutee PROPERTIES SOVERSION 1)
 set_target_properties(cutee PROPERTIES PUBLIC_HEADER 
 "\
 include/cutee/osutil.hpp;\
+include/cutee/function.hpp;\
 include/cutee/float_eq.hpp;\
 include/cutee/performance_test.hpp;\
 include/cutee/collection.hpp;\
