+++ include/cutee.hpp
@@ -6,14 +6,17 @@
 #endif /* CUTEE_OSTREAM_UTILITY */
 
 //#include "cutee/asserter.hpp"
+// Utility
 #include "cutee/exceptions.hpp"
 #include "cutee/timer.hpp"
 #include "cutee/float_eq.hpp"
+#include "cutee/function.hpp"
 
+// Interface
 #include "cutee/collection.hpp"
+#include "cutee/test_suite.hpp"
 
 #include "cutee/performance_test.hpp"
-#include "cutee/test_suite.hpp"
 #include "cutee/macros.hpp"
 #include "cutee/unit_assert.hpp"
 #include "cutee/unit_test_factory.hpp"
